Big refactor of internal state
Instead of a dozen lists for various states, have a single coroutine list and a CoroutineData object holding all the state. This may need some improvement later for performance, but we will drive that process via profiling. This makes the logic of things like pause/resume much simpler. The emulation test is passing again, and have fixed a couple of issues with pause/resume that are now far more obvious.