Fix for https://github.com/treething77/coroutinery/issues/1
Fixing an issue where we would get a range exception when processing source files outside of the Assets folder. We now take the full url in this case. I had thought this would cause an issue with OpenFileAtLineExternal because according to what I'd read it required the path relative to the Assets folder. This does not seem to be the case though. So now for source files outside of Assets you will see the full path in the debugger, but it should still work to click on the path to open the source file.